refactor(recipe): move list to component
